+++ apps/client/src/app/core/alarms/+state/alarms.facade.ts
@@ -268,8 +268,8 @@ export class AlarmsFacade {
     }
     let spawn = nextSpawn.hours;
     let despawn = nextSpawn.despawn;
-    despawn = despawn === 0 ? 24 : despawn;
-    spawn = spawn === 0 ? 24 : spawn;
+    despawn = despawn || 24;
+    spawn = spawn || 24;
     // If spawn is greater than despawn, it means that it spawns before midnight and despawns after, which is during the next day.
     const despawnsNextDay = spawn > despawn;
     if (!despawnsNextDay) {
@@ -349,16 +349,21 @@ export class AlarmsFacade {
         if (range) {
           const intersectSpawn = range[0];
           const intersectDespawn = range[1] || 24;
-          if (intersectSpawn < intersectDespawn) {
-            const days = Math.max(Math.floor((weatherSpawn.spawn.getTime() - time) / 86400000), 0);
-            return {
-              hours: intersectSpawn,
-              days: days,
-              despawn: intersectDespawn,
-              weather: weatherSpawn.weather,
-              date: weatherSpawn.spawn
-            };
-          }
+          // Set the snapshot date to the moment at which the node will spawn for real.
+          const date = weatherSpawn.spawn;
+          date.setUTCHours(intersectSpawn);
+          date.setUTCMinutes(0);
+          date.setUTCSeconds(0);
+          date.setUTCMilliseconds(0);
+          // Adding 3 seconds margin for days computation
+          const days = Math.max(Math.floor((weatherSpawn.spawn.getTime() - time + 3000 * EorzeanTimeService.EPOCH_TIME_FACTOR) / 86400000), 0);
+          return {
+            hours: intersectSpawn,
+            days: days,
+            despawn: intersectDespawn,
+            weather: weatherSpawn.weather,
+            date
+          };
         }
       }
     }