+++ apps/client/src/app/core/eorzea/timer.pipe.ts
@@ -11,10 +11,6 @@ export class TimerPipe implements PipeTransform {
   }
 
   transform(duration: number, verbose = false): string {
-    // 259200 is 3 days, we'll display a specific timer for that because of weather transition accuracy issues
-    // if (duration > 259200) {
-    //   return this.translate.instant('ALARMS.More_than_three_days');
-    // }
     const seconds = duration % 60;
     const minutes = Math.floor(duration / 60) % 60;
     const hours = Math.floor(duration / 3600) % 24;