+++ apps/client/src/environments/game-env.ts
@@ -1,6 +1,6 @@
 export const gameEnv = {
   maxLevel: 80,
-  globalGameVersion: 5.5,
+  globalGameVersion: 5.55,
   koreanGameVersion: 5.4,
   chineseGameVersion: 5.41
 };