+++ apps/client/src/environments/patch-notes.ts
@@ -1,4 +1,10 @@
 export const patchNotes = `### Bug Fixes
 
-* **alarms:** fixed an issue with some fish under specific conditions having wrong timers.
-* **pcap:** force powershell to stop.`;
+* **alarms:** better days computing system.
+* **alarms:** fixed an issue with weather + time coordination in some timers.
+* **alarms:** fixed some timers skipping a day and jumping from "spawn" to "tomorrow" after that.
+
+
+### Features
+
+* **db:** 5.55 content update.`;